requirements.txt
Django==4.2.7
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.17.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
joblib==1.3.2
tensorflow==2.13.0
reportlab==4.0.7
arabic-reshaper==3.0.0
python-bidi==0.4.2